+++ Sources/PippinAdapters/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Added
+
+- Allow multiple selections in `CrudViewController`.
+
 ## [5.1.0] 2020-05-28
 
 ### Added